+++ instagram-ts/source/ui/components/image/HalfBlock.tsx
